% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{getCellMeans}
\alias{getCellMeans}
\title{Get cell means for unique groups of covariates}
\usage{
getCellMeans(x, y, factor.groups)
}
\arguments{
\item{x}{a matrix-like object containing the covariates.}

\item{y}{a vector of values to compute the means from.}

\item{factor.groups}{a factor of unique groups of covariates.}
}
\value{
the cell means.
}
\description{
Get cell means for unique groups of covariates
}
