% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{getCellSizes}
\alias{getCellSizes}
\title{Get number of units in each cell}
\usage{
getCellSizes(x, n, factor.groups)
}
\arguments{
\item{x}{a matrix-like object containing the covariates.}

\item{n}{the total number of units.}

\item{factor.groups}{a factor of unique groups of covariates.}
}
\value{
the number of units in each cell.
}
\description{
Get number of units in each cell
}
