% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRRparameters.R
\name{getRRparameters}
\alias{getRRparameters}
\title{Compute Randomized Response parameters}
\usage{
getRRparameters(vec.RRmodel, vec.p1, vec.p2)
}
\arguments{
\item{vec.RRmodel}{a character vector of Randomized Response models.}

\item{vec.p1}{a numeric vector of p1 values.}

\item{vec.p2}{a numeric vector of p2 values.}
}
\value{
a list with c and d values.
}
\description{
Compute Randomized Response parameters
}
