% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{getUniqueGroups}
\alias{getUniqueGroups}
\title{Get unique groups of covariates}
\usage{
getUniqueGroups(x)
}
\arguments{
\item{x}{a matrix-like object containing the covariates.}
}
\value{
a factor of unique groups.
}
\description{
Get unique groups of covariates
}
