% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{intDotplot}
\alias{intDotplot}
\title{Internal function to draw a dotplot}
\usage{
intDotplot(x, mtit = "95\%-CI", ...)
}
\arguments{
\item{x}{data.}

\item{mtit}{main title.}
}
\description{
Sligthly modified function dotplot.ranef.mer from lme4 package.
}
\keyword{internal}
