% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.RRglmGOF}
\alias{print.RRglmGOF}
\title{Print RRglmGOF values}
\usage{
\method{print}{RRglmGOF}(x, digits = 3, ...)
}
\arguments{
\item{x}{an object of class RRglmGOF.}

\item{digits}{minimal number of \emph{significant digits} (default: 3).}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print RRglmGOF values
}
