% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.summary.RRglm}
\alias{print.summary.RRglm}
\title{Print RRglm summary}
\usage{
\method{print}{summary.RRglm}(
  x,
  printPrevalence = TRUE,
  printPrevalencePerLevel = FALSE,
  printResiduals = FALSE,
  digits = 5,
  ...
)
}
\arguments{
\item{x}{an object of class summary.RRglm.}

\item{printPrevalence}{print estimated population prevalence per item and RR model (default: true).}

\item{printPrevalencePerLevel}{print estimated population prevalence per item, RRmodel and protection level (default: false).}

\item{printResiduals}{print deviance residuals (default: false).}

\item{digits}{minimal number of \emph{significant digits} (default: 5).}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print RRglm summary
}
