% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{residuals.RRglmerMod}
\alias{residuals.RRglmerMod}
\title{Accessing GLMMRR Fits for mixed-effect models}
\usage{
\method{residuals}{RRglmerMod}(
  object,
  type = c("deviance", "pearson", "working", "response", "partial",
    "unconditional.response", "unconditional.pearson"),
  ...
)
}
\arguments{
\item{object}{an object of class RRglmer.}

\item{type}{the type of residuals which should be returned. The alternatives are: "deviance" (default), "pearson", "working", "response", "partial",
"unconditional.response" and "unconditional.pearson".}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A vector of residuals.
}
\description{
Compute residuals for RRglmer objects. Extends \code{\link{residuals.glmResp}} to access conditional and
unconditional residuals for grouped binary Randomized Response data.
}
\seealso{
\code{\link{residuals.glmResp}}
}
