% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRbinomial.R
\name{RRbinomial}
\alias{RRbinomial}
\title{Binomial family adjusted for Randomized Response parameters.}
\usage{
RRbinomial(link, c, d, ...)
}
\arguments{
\item{link}{a specification for the model link function. Must be an object of class "link-glm".}

\item{c}{a numeric vector containing the parameter c.}

\item{d}{a numeric vector containing the parameter d.}

\item{...}{other potential arguments to be passed to \code{\link{binomial}}.}
}
\value{
A binomial family object.
}
\description{
The upper and lower limits for mu's depend on the Randomized Response parameters.
}
\seealso{
\code{\link{family}}
}
