% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRlink_cauchit.R
\name{RRlink.cauchit}
\alias{RRlink.cauchit}
\title{Cauchit link function with Randomized Response parameters.}
\usage{
RRlink.cauchit(c, d)
}
\arguments{
\item{c}{a numeric vector containing the parameter c.}

\item{d}{a numeric vector containing the parameter d.}
}
\value{
RR link function.
}
\description{
Cauchit link function with Randomized Response parameters.
}
