% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRlink_cloglog.R
\name{RRlink.cloglog}
\alias{RRlink.cloglog}
\title{Log-Log link function with Randomized Response parameters.}
\usage{
RRlink.cloglog(c, d)
}
\arguments{
\item{c}{a numeric vector containing the parameter c.}

\item{d}{a numeric vector containing the parameter d.}
}
\value{
RR link function.
}
\description{
Log-Log link function with Randomized Response parameters.
}
