% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.RRglm}
\alias{summary.RRglm}
\title{Summarizing GLMMRR Fits for fixed-effect models}
\usage{
\method{summary}{RRglm}(object, printResiduals = FALSE,
  limitRRparameters = 10, digits = 5, ...)
}
\arguments{
\item{object}{an object of class RRglm.}

\item{printResiduals}{print scaled Pearson residuals (default: false).}

\item{limitRRparameters}{set limit for list of Randomized Response parameters (default: 10).}

\item{digits}{minimal number of \emph{significant digits} (default: 5).}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Summarizing GLMMRR Fits for fixed-effect models
}

