% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMLPrevalence.R
\name{getMLPrevalence}
\alias{getMLPrevalence}
\title{Compute Estimated Population Prevalence}
\usage{
getMLPrevalence(mu, n, c, d)
}
\arguments{
\item{mu}{observed mean response.}

\item{n}{number of units.}

\item{c}{randomized response parameter c.}

\item{d}{randomized response parameter d.}
}
\value{
maximum likelihood estimate of the population prevalence and its variance.
}
\description{
Compute Estimated Population Prevalence
}
