% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRglmGOF.R
\name{RRglmGOF}
\alias{RRglmGOF}
\title{Goodness-of-fit statistics for binary Randomized Response data}
\usage{
RRglmGOF(
  RRglmOutput,
  doPearson = TRUE,
  doDeviance = TRUE,
  doHlemeshow = TRUE,
  hlemeshowGroups = 10,
  rm.na = TRUE
)
}
\arguments{
\item{RRglmOutput}{a model fitted with the \code{\link{RRglm}} function.}

\item{doPearson}{compute Pearson statistic.}

\item{doDeviance}{compute Deviance statistic.}

\item{doHlemeshow}{compute Hosmer-Lemeshow statistic.}

\item{hlemeshowGroups}{number of groups to split the data into for the Hosmer-Lemeshow statistic (default: 10).}

\item{rm.na}{remove cases with missing data.}
}
\value{
an option of class RRglmGOF.
}
\description{
Compute goodness-of-fit statistics for binary Randomized Response data.
Pearson, Deviance and Hosmer-Lemeshow statistics are available.
}
\examples{
out <- RRglm(response ~ Gender + RR + pp + age, link="RRlink.logit", RRmodel=RRmodel,
         p1=RRp1, p2=RRp2, data=Plagiarism, etastart=rep(0.01, nrow(Plagiarism)))
RRglmGOF(RRglmOutput = out, doPearson = TRUE, doDeviance = TRUE, doHlemeshow = TRUE)
}
