% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRlink_probit.R
\name{RRlink.probit}
\alias{RRlink.probit}
\title{Probit link function with Randomized Response parameters.}
\usage{
RRlink.probit(c, d)
}
\arguments{
\item{c}{a numeric vector containing the parameter c.}

\item{d}{a numeric vector containing the parameter d.}
}
\value{
RR link function.
}
\description{
Reference: Fox, J-P, Klotzke, K. and Veen, D. (2016). \emph{Generalized Linear Mixed Models for Randomized
Responses.} Manuscript submitted for publication.
}

